fireflies/requirements.txt
python-dotenv
requests>=2.28.0
